+++ test/system-ca/system-ca.status
@@ -1,0 +1,7 @@
+prefix system-ca
+
+# To mark a test as flaky, list the test name in the appropriate section
+# below, without ".js", followed by ": PASS,FLAKY". Example:
+# sample-test                        : PASS,FLAKY
+
+[true] # This section applies to all platforms
